moving GetLitmusChaosMetrics func to controller
